# Conflicts:
#	Melin.Server/Controllers/AuthController.cs

+++ Melin.Server/Controllers/ReferenceController.cs
@@ -1,24 +1,15 @@
-﻿using System.Text.Json;
-using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Threading.Tasks;
+﻿using Microsoft.AspNetCore.Mvc;
+using System.Globalization;
 using Melin.Server.Filter;
-using Melin.Server.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Melin.Server.Models;
-using Melin.Server.Models.Context;
-using Melin.Server.Models.References;
+using Melin.Server.Models.DTO;
 using Melin.Server.Services;
 using Melin.Server.Wrappers;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.JsonPatch;
-using Microsoft.EntityFrameworkCore;
-using Presentation = Melin.Server.Models.Presentation;
-using Report = Melin.Server.Models.Report;
-using Software = Melin.Server.Models.Software;
-using Task = System.Threading.Tasks.Task;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
+using Serilog;
 
 namespace Melin.Server.Controllers;
 
@@ -37,47 +28,125 @@ public class ReferenceController : ControllerBase
         _tagService = tagService;
     }
 
+    /// <summary>
+    /// Retrieves a single Reference from the currently logged-in user
+    /// </summary>
+    /// <param name="refId">Integer Query Parameter referring to the ID of the Reference</param>
+    /// <returns>A Single Reference</returns>
     [HttpGet("get-single-reference")]
     [Authorize]
-    public async Task<IActionResult> GetSingleReference(int refId)
+    [ProducesResponseType(typeof(ICollection<Reference>), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    public async Task<IActionResult> GetSingleReference([FromQuery] int refId)
     {
         try
         {
+            if (User.Identity == null)
+            {
+                Log.Information("Unauthorized Attempt to Retrieve Reference: {ReferenceID}", refId);
+                return Unauthorized();
+            }
+            Log.Information("GET: SingleReference of ID: {refId}", refId);
             var reference = await _referenceService.GetReferenceWithAllDetailsById(User.Identity.Name, refId);
-
+            
             if (reference.Success)
             {
+                Log.Information("GET: Reference ID: {refId} Successfully retrieved by User: {userEmail}", refId, User.Identity.Name);
                 return Ok(reference.Data);
             }
-            else
-            {
-                return NotFound("REFERENCE NOT FOUND");
-            }
+
+            Log.Information("GET: Reference Not Found ID: {refId}", refId);
+            return NotFound("REFERENCE NOT FOUND");
         }
         catch (Exception e)
         {
-            Console.WriteLine(e);
+            Log.Error("GET: Unable to retrieve single reference of id: {refId}", refId);
             return BadRequest();
         }
     }
     
+    
+    /// <summary>
+    /// Retrieves a list of References from the currently logged-in user
+    /// </summary>
+    /// <param name="filter">A <see cref="PaginationFilter"/> object. Determines page number and quantity of References</param>
+    /// <returns>A 'filtered' <see cref="PagedResponse{T}"/></returns>
     [HttpGet("references")]
     [Authorize]
+    [ProducesResponseType(typeof(ICollection<Reference>), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
     public async Task<IActionResult> GetReferences([FromQuery] PaginationFilter filter)
     {
-        var userEmail = User.Identity.Name;
-        var validFilter = new PaginationFilter(filter.PageNumber, filter.PageSize);
-        validFilter.PageSize = 1000;
-        var pagedReferences = await _referenceService.GetOwnedReferencesAsync(filter, userEmail);
-
-        var totalRefCount = await _referenceService.GetOwnedReferenceCountAsync(userEmail);
+        try
+        {
+            if (User.Identity == null)
+            {
+                Log.Information("Unauthorized Attempt to Retrieve Multiple References");
+                return Unauthorized();
+            }
+            var userEmail = User.Identity.Name;
+            var validFilter = new PaginationFilter(filter.PageNumber, filter.PageSize)
+            {
+                PageSize = 1000
+            };
         
-        return Ok(new PagedResponse<ICollection<Reference>>(pagedReferences, validFilter.PageNumber, validFilter.PageSize, totalRefCount));
+            var pagedReferences = await _referenceService.GetOwnedReferencesAsync(filter, userEmail);
+
+            var totalRefCount = await _referenceService.GetOwnedReferenceCountAsync(userEmail);
+
+            if (totalRefCount == 0)
+            {
+                Log.Information("GET: {userEmail}, attempted GetReferences, -> currently has 0",  User.Identity.Name);
+                return NotFound("User currently has 0 references");
+            }
+            Log.Information("{userEmail}, retrieved {referenceAmount} References", User.Identity.Name, totalRefCount);
+            
+            // TEMP CONVERSION TESTING OUT DTO METHOD
+            List<ReferenceToLibraryRequest> output = new List<ReferenceToLibraryRequest>();
+            foreach (var reference in pagedReferences)
+            {
+                var res = new ReferenceToLibraryRequest
+                {
+                    Id = reference.Id,
+                    Type = reference.Type.ToString(),
+                    Title = reference.Title,
+                    CreatedAt = reference.CreatedAt.ToString(CultureInfo.CurrentCulture),
+                    UpdatedAt = reference.UpdatedAt.ToString(CultureInfo.CurrentCulture),
+                    Creators = reference.Creators?.ToList(),
+                    Tags = reference.Tags?.ToList(),
+                    Language = reference.Language.ToString(),
+                    DatePublished = reference.DatePublished.ToString()
+                };
+                output.Add(res);
+            }
+
+            return Ok(new PagedResponse<ICollection<ReferenceToLibraryRequest>>(output, validFilter.PageNumber, validFilter.PageSize, totalRefCount));
+        }
+        catch (Exception e)
+        {
+            Log.Error("GET: GetReferences Failed. PaginationFilter: {filter}. By User: {userEmail}", filter, User.Identity.Name);
+            return BadRequest();
+        }
+
     }
 
+    /// <summary>
+    /// Creates a new Reference and sets the initial owner to the current User.
+    /// Uses a custom JSON converter to handle the 30+ types of References on a single
+    /// endpoint
+    /// </summary>
+    /// <param name="reference">A <see cref="Reference"/></param>
+    /// <returns><see cref="StatusCodeResult"/></returns>
     // POST: Create new reference
     [HttpPost("create-reference")]
     [Authorize]
+    [ProducesResponseType(typeof(string), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public async Task<ActionResult> PostReference([FromBody] Reference reference) {
         if (!ModelState.IsValid)
         {
@@ -86,76 +155,51 @@ public class ReferenceController : ControllerBase
         
         try
         {
-            if (User.Identity.Name == null)
+            if (User.Identity == null)
             {
+                Log.Warning("Unauthorized access attempted");
                 return Unauthorized("User not authorized to create References");
             }
             reference.OwnerEmail = User.Identity.Name;
 
             await _referenceService.AddReferenceAsync(reference);
-            return Ok("Artwork created successfully");
+            Log.Information("{userEmail} Created a Reference with ID: {refId}",  User.Identity.Name, reference.Id);
+
+            return Ok("Reference created successfully");
         } catch (Exception ex) {
-            return StatusCode(500, "An error occurred while creating the reference.");
+            Log.Warning("Unable to create reference");
+            return BadRequest();
         }
     }
     
-    private async Task<bool> HandleTagsWithReferencePost(ICollection<Tag> tags)
-    {
-        try
-        {
-            if (tags.Count > 1)
-            {
-                string createdBy = User.Identity.Name;
-                await _tagService.CreateTagsAsync(tags, createdBy);
-            }
-            else
-            {
-                foreach (var tag in tags)
-                {
-                    tag.CreatedBy = User.Identity.Name;
-                    var existingTag = await _tagService.GetTagAsync(tag.Id);
-                    if (existingTag == null)
-                    {
-                        await _tagService.CreateTagAsync(tag);
-                    }
-                }
-            }
-
-            return true;
-        }
-        catch (Exception e)
-        {
-            Console.WriteLine(e);
-            return false;
-        }
-
-    }
-    
+    /// <summary>
+    /// Updates a User owned Reference, via HTTP PATCH.
+    /// </summary>
+    /// <param name="id">Integer value depicting the Reference to be updated</param>
+    /// <param name="updatedItem"><see cref="JsonPatchDocument"/></param>
+    /// <returns><see cref="IActionResult"/></returns>
     // UPDATE
     [HttpPatch("update/{id}")]
     [Authorize]
-    public async Task<IActionResult> UpdateItem(int id, [FromBody] JsonPatchDocument<Reference> updatedItem)
+    [ProducesResponseType(typeof(IActionResult), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    public async Task<IActionResult> UpdateItem([FromQuery] int id, [FromBody] JsonPatchDocument<Reference> updatedItem)
     {
-        // Validate model
         if (!ModelState.IsValid)
         {
             return BadRequest(ModelState);
         }
 
-        if (updatedItem == null)
+        if (User.Identity?.Name == null)
         {
-            return BadRequest(ModelState);
-        }
-
-        if (User.Identity == null)
-        {
+            Log.Information("Unauthorized attempt to update Reference: {RefID}", id);
             return Unauthorized("User Currently Not Authorized to Update Item");
         }
-
-        if (User.Identity.Name == null)
-        {
-            return Unauthorized("User Has Incorrect Auth Details. Attempt to Login Again: Please Advise.");
-        }
+        
+        Log.Information("PATCH: {userEmail}, updating reference", User.Identity.Name);
+        
 
         // Find the item to update
         var existingItemResult = await _referenceService.GetReferenceWithAllDetailsById(User.Identity.Name, id);
@@ -165,6 +209,11 @@ public class ReferenceController : ControllerBase
         }
 
         var existingItem = existingItemResult.Data;
+
+        if (existingItem == null)
+        {
+            return BadRequest();
+        }
         
         updatedItem.ApplyTo(existingItem, ModelState);
 
@@ -176,21 +225,39 @@ public class ReferenceController : ControllerBase
         // apply patch to DB
         _referenceService.ApplyPatch(existingItem);
         return new ObjectResult(existingItem);
-        
     }
 
-
+    /// <summary>
+    /// Deletes a single Reference
+    /// </summary>
+    /// <param name="refId">Integer Query Parameter for Reference ID</param>
+    /// <returns><see cref="ActionResult"/></returns>
     // DELETE: Delete single reference
     [HttpDelete("delete-reference")]
     [Authorize]
-    public async Task<ActionResult<bool>> DeleteSpecificReference(int refId)
+    [ProducesResponseType(typeof(ActionResult<bool>), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    public async Task<ActionResult<bool>> DeleteSpecificReference([FromQuery] int refId)
     {
         try
         {
+            if (User.Identity == null)
+            {
+                Log.Information("Unauthorized Attempt to delete Reference ID: {ReferenceID}", refId);
+                return Unauthorized();
+            }
+            
+            if (User.Identity.Name == null)
+            {
+                Log.Information("Email is Null when attempting to delete Reference ID: {ReferenceID}", refId);
+                return Unauthorized();
+            }
 
             var r = await _referenceService.DeleteReferenceAsync(User.Identity.Name, refId);
             
-            if (r == null)
+            if (!r)
             {
                 return NotFound("Reference with ID: " + refId + " not found.");
             }
@@ -200,21 +267,42 @@ public class ReferenceController : ControllerBase
         }
         catch (Exception e)
         {
-            Console.WriteLine(e);
-            throw;
+            Log.Warning("Unable to Delete Reference with ID: {ReferenceID}", refId);
+            return BadRequest();
         }
     }
     
+    /// <summary>
+    /// Deletes a List of References
+    /// </summary>
+    /// <param name="refIdList">A List of Integer's depicting Reference ID's</param>
+    /// <returns></returns>
     // DELETE: Delete list of references
     [HttpDelete("delete-multiple-references")]
     [Authorize]
+    [ProducesResponseType(typeof(ActionResult<bool>), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public async Task<ActionResult<bool>> DeleteListOfReferences(List<int> refIdList)
     {
         try
         {
             if (refIdList.Count < 1)
             {
+                Log.Warning("Attempted to Delete from a List that has no References");
                 return Problem("ERROR: Ref ID List has < 1 ids");
+            }
+
+            if (User.Identity == null)
+            {
+                Log.Information("Unauthorized Attempt to delete Reference ID: {ReferenceID}", refIdList);
+                return Unauthorized();
+            }
+            
+            if (User.Identity.Name == null)
+            {
+                Log.Information("Email is Null when attempting to delete Reference ID: {ReferenceID}", refIdList);
+                return Unauthorized();
             }
 
             await _referenceService.DeleteReferenceRangeAsync(User.Identity.Name, refIdList);
@@ -225,8 +313,8 @@ public class ReferenceController : ControllerBase
         }
         catch (Exception e)
         {
-            Console.WriteLine(e);
-            throw;
+            Log.Warning("Attempt to Delete list of References failed. ReferenceIDs: {ReferenceIDList}", refIdList);
+            return BadRequest();
         }
     }
 }
